+++ react-app/src/library/types/consumed-headset-events.ts
@@ -11,7 +11,6 @@ export enum HeadsetEvents {
     loggableEvent = 'loggableEvent',
     webHidPermissionRequested = 'webHidPermissionRequested',
     deviceConnectionStatusChanged = 'deviceConnectionStatusChanged',
-    deviceMuteChanged = 'deviceMuteChanged',
     deviceEventLogs = 'deviceEventLogs',
 }